Add new data structure heapq

+++ src/rr_array.h
@@ -2,23 +2,22 @@
 #define _RR_ARRAY_H
 
 #include <stddef.h>
-#include <stdint.h>
 
 typedef struct array_t {
-    uint32_t    nelm;        /* current # of elements in array   */
-    uint32_t    nrest;       /* rest # of elements in array      */
-    size_t      size;        /* size of each item in array       */
-    void        *elm;        /* point to the first item in array */
+    unsigned long    nelm;        /* current # of elements in array   */
+    unsigned long    nrest;       /* rest # of elements in array      */
+    size_t           size;        /* size of each item in array       */
+    void             *elm;        /* point to the first item in array */
 } array_t;
 
 #define ARRAY_AT(a, i) ((void *)(((char *)(a)->elm + (a)->size*(i))))
 #define ARRAY_LEN(a) ((a)->nelm)
 
-array_t * array_create(uint32_t n, size_t s);
+array_t * array_create(unsigned long n, size_t s);
 void array_free(array_t *array);
 void * array_push(array_t * array);
-void * array_push_n(array_t * array, uint32_t n);
-uint32_t array_len(array_t * array);
-void * array_at(array_t * array, int i);
+void * array_push_n(array_t * array, unsigned long n);
+unsigned long array_len(array_t * array);
+void * array_at(array_t * array, long i);
 
 #endif
